fix: delete img onarticle

+++ src/controller/articleController.php
@@ -47,15 +47,17 @@ class articleController
         if(isset($_POST['deleteUser'])) {
             $id = $_POST['education_id_delete'];
             $education = $this->educationRepo->getById($id);
-            $img = $education->getImg();
+            if($education->getImg() != null){
+                $img = $education->getImg();
 
-            $destination_path = getcwd() . DIRECTORY_SEPARATOR . 'assetsWeb' . DIRECTORY_SEPARATOR . 'img' . DIRECTORY_SEPARATOR . 'education' . DIRECTORY_SEPARATOR;
-            $targetFile = $destination_path . basename($img);
+                $destination_path = getcwd() . DIRECTORY_SEPARATOR . 'assetsWeb' . DIRECTORY_SEPARATOR . 'img' . DIRECTORY_SEPARATOR . 'education' . DIRECTORY_SEPARATOR;
+                $targetFile = $destination_path . basename($img);
 
-            if(file_exists($targetFile)){
-                unlink($targetFile);
-
+                if(file_exists($targetFile)){
+                    unlink($targetFile);
+                }
             }
+
             $this->educationRepo->delete($id);
         }
         $user = $this->sessionService->current();
